Move duplicate code into superclass
