+++ app/src/org/commcare/activities/MenuList.java
@@ -1,10 +1,6 @@
 package org.commcare.activities;
 
-import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.AdapterView.OnItemClickListener;
 import android.widget.ListView;
 import android.widget.TextView;
 
@@ -12,15 +8,11 @@ import org.commcare.CommCareApplication;
 import org.commcare.adapters.MenuAdapter;
 import org.commcare.dalvik.R;
 import org.commcare.fragments.BreadcrumbBarFragment;
-import org.commcare.session.SessionFrame;
-import org.commcare.suite.model.Entry;
-import org.commcare.suite.model.Menu;
-import org.commcare.util.CommCarePlatform;
 import org.commcare.views.ManagedUi;
 import org.commcare.views.UiElement;
 
 @ManagedUi(R.layout.screen_suite_menu)
-public class MenuList extends SaveSessionCommCareActivity implements OnItemClickListener {
+public class MenuList extends MenuBase {
 
     private MenuAdapter adapter;
 
@@ -30,18 +22,9 @@ public class MenuList extends SaveSessionCommCareActivity implements OnItemClick
     // removed the UiElement annotation here because it was causing a crash @ loadFields() in CommCareActivity
     private TextView header;
 
-    private boolean isRootModuleMenu;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        CommCarePlatform platform = CommCareApplication._().getCommCarePlatform();
-        String menuId = getIntent().getStringExtra(SessionFrame.STATE_COMMAND_ID);
-
-        if (menuId == null) {
-            menuId = Menu.ROOT_MENU_ID;
-            isRootModuleMenu = true;
-        }
 
         if (header == null) {
             header = (TextView)getLayoutInflater().inflate(R.layout.menu_list_header, null);
@@ -53,27 +36,11 @@ public class MenuList extends SaveSessionCommCareActivity implements OnItemClick
             list.addHeaderView(header, null, false);
         }
 
-        adapter = new MenuAdapter(this, platform, menuId);
+        adapter = new MenuAdapter(this, CommCareApplication._().getCommCarePlatform(), menuId);
         adapter.showAnyLoadErrors(this);
         refreshView();
 
         list.setOnItemClickListener(this);
-    }
-
-    @Override
-    protected boolean isTopNavEnabled() {
-        return true;
-    }
-
-    @Override
-    public String getActivityTitle() {
-        //return adapter.getMenuTitle();
-        return null;
-    }
-
-    @Override
-    public boolean isBackEnabled() {
-        return !(CommCareApplication._().isConsumerApp() && isRootModuleMenu);
     }
 
     /**
@@ -82,35 +49,4 @@ public class MenuList extends SaveSessionCommCareActivity implements OnItemClick
     private void refreshView() {
         list.setAdapter(adapter);
     }
-
-    /**
-     * Stores the path of selected form and finishes.
-     */
-    @Override
-    public void onItemClick(AdapterView listView, View view, int position, long id) {
-        String commandId;
-        Object value = listView.getAdapter().getItem(position);
-        // if value is null, probably it means that we clicked on the header view, so we just ignore it
-        if (value == null) {
-            return;
-        }
-        if (value instanceof Entry) {
-            commandId = ((Entry)value).getCommandId();
-        } else {
-            commandId = ((Menu)value).getId();
-        }
-
-        // create intent for return and store path
-        Intent i = new Intent(getIntent());
-        i.putExtra(SessionFrame.STATE_COMMAND_ID, commandId);
-        setResult(RESULT_OK, i);
-
-        finish();
-    }
-
-    protected boolean onBackwardSwipe() {
-        onBackPressed();
-        return true;
-    }
-
 }
